requirements.txt
discord
Pillow
psutil
requests